Add ability to set tcpdf header and footer data
By passing feader and footer data in PdfMerge we can directly
influence how and what data will be present while merging pdf
files.

The reason for implementing this feature in our case is that we want
to get rid of lines in headers and footers, so we can set line color
to white color.